feat: simplify k8s client by removing abstraction
