+++ crates/provider/src/kubernetes.rs
@@ -3,5 +3,5 @@ mod namespace;
 mod node;
 mod provider;
 
-pub use client::{KubeRsKubernetesClient, KubernetesClient};
+pub use client::KubernetesClient;
 pub use provider::KubernetesProvider;
