+++ crates/sdk/src/lib.rs
@@ -4,7 +4,7 @@ pub use orchestrator::{
     errors::OrchestratorError, network::Network, AddCollatorOptions, AddNodeOptions, Orchestrator,
     PjsResult,
 };
-use provider::{KubeRsKubernetesClient, KubernetesProvider, NativeProvider, ProviderError};
+use provider::{KubeRsClient, KubernetesProvider, NativeProvider, ProviderError};
 use support::{fs::local::LocalFileSystem, process::os::OsProcessManager};
 
 #[async_trait]
@@ -34,10 +34,7 @@ impl NetworkConfigExt for NetworkConfig {
     }
 
     async fn spawn_k8s(self) -> Result<Network<LocalFileSystem>, OrchestratorError> {
-        let client = KubeRsKubernetesClient::new(LocalFileSystem {})
-            .await
-            .map_err(|e| ProviderError::CreateClientFailed(String::from("Kubernetes"), e.into()))?;
-        let provider = KubernetesProvider::new(LocalFileSystem {}, client);
+        let provider = KubernetesProvider::new(LocalFileSystem {}).await;
         let orchestrator = Orchestrator::new(LocalFileSystem {}, provider);
         orchestrator.spawn(self).await
     }
